% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beers.R
\name{Beers}
\alias{Beers}
\title{Beers multipliers}
\usage{
Beers(data, ...)
}
\arguments{
\item{data}{data.drame. It contains at least two variables: five-year grouped ages and population.}

\item{...}{Arguments to be passed to \code{dplyr::select}, i.e., age and population, respectively.}
}
\value{
\code{Beers} returns a data.frame with specific ages and populations.
}
\description{
Method to open five-year grouped ages into specific ages.
}
\examples{

Beers(Ecuador1990, age, population)


}
\references{
\insertRef{Shryock}{popstudy}
}
\seealso{
\code{\link{Sprague}}
}
\author{
Cesar Gamboa-Sanabria
}
