\name{CEB}
\alias{CEB}
\docType{data}
\title{
Children Ever Born Data
}
\description{
Children Ever Born Data from Bolivia's 2001 Census data.
}
\usage{data("CEB")}
\format{
  A data frame with 27 observations on 8 variables for each five-year grouped age.
}
\source{
\href{https://www.ine.gob.bo/}{https://www.ine.gob.bo/}
}
\examples{
data(CEB)
summary(CEB)
}
\keyword{datasets}
