\name{CR_fertility_rates_1950_2011}
\alias{CR_fertility_rates_1950_2011}
\docType{data}
\title{
Costa Rica fertility rates
}
\description{
Fertility rates for Costa Rica 1950-2011.
}
\usage{data("CR_fertility_rates_1950_2011")}
\format{
  A data frame with 2170 observations on the following 3 variables.
  \describe{
    \item{\code{Year}}{a numeric vector}
    \item{\code{Age}}{a numeric vector}
    \item{\code{Female}}{a numeric vector with fertility rates}
  }
}
\source{
\href{https://www.inec.go.cr/}{https://www.inec.go.cr/}
}
\examples{
data(CR_fertility_rates_1950_2011)
summary(CR_fertility_rates_1950_2011)
}
\keyword{datasets}

