% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karup_king.R
\name{karup_king}
\alias{karup_king}
\title{karup_king}
\usage{
karup_king(data)
}
\arguments{
\item{data}{data.frame. A dataset with two variables: \code{age}, the group age each 5 years; and \code{pop}, the population for that age.}
}
\value{
\code{karup_king} function returns a a data frame with separated simple ages.
}
\description{
Separate grouped-age data to simple ages data using Karup-King separation factors.
}
\examples{

karup_king(grouped_age_CR_pop)

}
\references{
\insertRef{karup_king}{popstudy}
}
\seealso{
\code{\link{grouped_age_CR_pop}}
}
\author{
Cesar Gamboa-Sanabria
}
