% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myers.R
\name{Myers}
\alias{Myers}
\title{Myer's Blended Index}
\usage{
Myers(data, ...)
}
\arguments{
\item{data}{data.drame. It contains at least two variables: specific ages and population.}

\item{...}{Arguments to be passed to \code{dplyr::select}, i.e., age and population, respectively.}
}
\value{
\code{Myers} returns a list with two objects:

\item{Mmat}{a data.frame with specific digits index}
\item{MI}{the Myer's Blend Index.}
}
\description{
An upgrade over the Whipple index allows analyzing digit's attraction (or repulsion) from 0 to 9.
}
\examples{

results <- Myers(Panama1990, age, pop)
results$Mmat
results$MI

}
\references{
\insertRef{Shryock}{popstudy}
}
\author{
Cesar Gamboa-Sanabria
}
