% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netmigration_projection.R
\name{netmigration_projection}
\alias{netmigration_projection}
\title{netmigration_projection}
\usage{
netmigration_projection(
  mortality_rates_path,
  TFR_path,
  total_population_path,
  WRA_path,
  omega_age,
  horizon,
  first_year_projection
)
}
\arguments{
\item{mortality_rates_path}{PENDIENTE}

\item{TFR_path}{PENDIENTE}

\item{total_population_path}{PENDIENTE}

\item{WRA_path}{PENDIENTE}

\item{omega_age}{PENDIENTE}

\item{horizon}{PENDIENTE}

\item{first_year_projection}{PENDIENTE}
}
\value{
\code{netmigration_projection} PENDIENTE
}
\description{
Forecasting net migration.
}
\examples{

## PENDIENTE

}
\author{
Cesar Gamboa-Sanabria
}
