% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op.arima.R
\name{op.arima}
\alias{op.arima}
\title{op.arima}
\usage{
op.arima(
  arima_process = c(p = 1, d = 1, q = 1, P = 1, D = 1, Q = 1),
  seasonal_periodicity,
  time_serie,
  reg = NULL,
  horiz = 12,
  prop = 0.8,
  training_weight = 0.2,
  testing_weight = 0.8,
  parallelize = FALSE,
  clusters = detectCores(logical = FALSE),
  ...
)
}
\arguments{
\item{arima_process}{PENDIENTE}

\item{seasonal_periodicity}{PENDIENTE}

\item{time_serie}{PENDIENTE}

\item{reg}{PENDIENTE}

\item{horiz}{PENDIENTE}

\item{prop}{PENDIENTE}

\item{training_weight}{PENDIENTE}

\item{testing_weight}{PENDIENTE}

\item{parallelize}{PENDIENTE}

\item{clusters}{PENDIENTE}

\item{...}{PENDIENTE}
}
\value{
\code{correlate_df} function returns a list with three objects: A data-frame with the correlation matrix and two correlation plots.
}
\description{
Estimates the best ARIMA model using overparameterization.
}
\details{
PENDIENTE
}
\examples{

#PENDIENTE

}
\references{

}
\author{
Cesar Gamboa-Sanabria
}
