\name{Panama1990}
\alias{Panama1990}
\docType{data}
\title{
Panama1990
}
\description{
Panama census data in 1990 by specific ages.
}
\usage{data("Panama1990")}
\format{
  A data frame with 100 observations on the following 2 variables.
  \describe{
    \item{\code{age}}{a character vector with specific ages}
    \item{\code{pop}}{a numeric vector with population for each age}
  }
}
\source{
\href{https://ccp.ucr.ac.cr/}{https://ccp.ucr.ac.cr/}
}
\examples{
data(Panama1990)
summary(Panama1990)
}
\keyword{datasets}
