% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popstudy.R
\docType{package}
\name{popstudy}
\alias{popstudy}
\title{popstudy Package}
\description{
Applied techniques to demographic and time series analysis.
}
\author{
Cesar Gamboa-Sanabria \email{info@cesargamboasanabria.com}
}
