% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_projection.R
\name{population_projection}
\alias{population_projection}
\title{population_projection}
\usage{
population_projection(...)
}
\arguments{
\item{...}{PENDIENTE}
}
\value{
\code{population_projection} PENDIENTE
}
\description{
Forecasting population
}
\examples{

## PENDIENTE

}
\author{
Cesar Gamboa-Sanabria
}
