% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_structure.R
\name{project_structure}
\alias{project_structure}
\title{project_structure}
\usage{
project_structure()
}
\value{
\code{project_structure} create basic diretories and files in the current working direcotory/repository.
}
\description{
Create a basic structure for a project repo.
}
\examples{
\donttest{
\dontrun{
project_structure()
}
}
}
\author{
Cesar Gamboa-Sanabria
}
