% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_from_dir.R
\name{read_from_dir}
\alias{read_from_dir}
\title{read_from_dir}
\usage{
read_from_dir(file, path = NULL)
}
\arguments{
\item{file}{The file name.}

\item{path}{The file location.}
}
\value{
\code{read_from_dir} Returns the normalizaed path for a file.
}
\description{
Get full path from a file.
}
\examples{
\donttest{
file.create("test_file.txt")
read_from_dir("test_file.txt")
}

}
\author{
Cesar Gamboa-Sanabria
}
