% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_packages.R
\name{required_packages}
\alias{required_packages}
\title{required_packages}
\usage{
required_packages(...)
}
\arguments{
\item{...}{packages names.}
}
\description{
Install/load the required packages from CRAN.
}
\examples{

required_packages(tidyr, dplyr, ggplot2)

}
\author{
Cesar Gamboa-Sanabria
}
