% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFR_projection.R
\name{TFR_projection}
\alias{TFR_projection}
\title{TFR_projection}
\usage{
TFR_projection(TFR_path, WRA_path, horizon, first_year_projection, ...)
}
\arguments{
\item{TFR_path}{PENDIENTE}

\item{WRA_path}{PENDIENTE}

\item{horizon}{PENDIENTE}

\item{first_year_projection}{PENDIENTE}

\item{...}{PENDIENTE}
}
\value{
\code{TFR_projection} PENDIENTE
}
\description{
Forecasting total fertility rates.
}
\examples{

## PENDIENTE

}
\author{
Cesar Gamboa-Sanabria
}
