% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_projection.R
\name{mortality_projection}
\alias{mortality_projection}
\title{mortality_projection}
\usage{
mortality_projection(
  mortality_rates_path,
  total_population_path,
  omega_age,
  horizon,
  first_year_projection,
  ...
)
}
\arguments{
\item{mortality_rates_path}{PENDIENTE}

\item{total_population_path}{PENDIENTE}

\item{omega_age}{PENDIENTE}

\item{horizon}{PENDIENTE}

\item{first_year_projection}{PENDIENTE}

\item{...}{PENDIENTE}
}
\value{
\code{mortality_projection} PENDIENTE
}
\description{
Forecasting mortality rates.
}
\examples{

## PENDIENTE

}
\author{
Cesar Gamboa-Sanabria
}
