% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_packages.R
\name{required_packages}
\alias{required_packages}
\title{required_packages}
\usage{
required_packages(...)
}
\arguments{
\item{...}{packages names.}
}
\value{
\code{required_packages} does not return a value, it only install and load the desired packages.
}
\description{
Install/load the required packages from CRAN.
}
\examples{
\donttest{
\dontrun{
#If you need to install and load the tidyr, dplyr and ggplot2 packages, run the following line:
#required_packages(tidyr, dplyr, ggplot2)
}
}
}
\author{
Cesar Gamboa-Sanabria
}
