% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sprague.R
\name{Sprague}
\alias{Sprague}
\title{Sprague multipliers}
\usage{
Sprague(data, ...)
}
\arguments{
\item{data}{data.drame. It contains at least two variables: five-year grouped ages and population.}

\item{...}{Arguments to be passed to \code{dplyr::select}, i.e., age and population, respectively.}
}
\value{
\code{Sprague} returns an object of class data.frame with population for specific ages.
}
\description{
Method to open five-year grouped ages into specific ages.
}
\examples{

Sprague(Ecuador1990, age, population)

}
\references{
\insertRef{Shryock}{popstudy}
}
\seealso{
\code{\link{Beers}}
}
\author{
Cesar Gamboa-Sanabria
}
